feat(file-upload-input): refactor components
