+++ src/sass/bundles/skin-headless.scss
@@ -9,7 +9,6 @@
 @import "../button/button";
 @import "../carousel/carousel";
 @import "../calendar/calendar";
-@import "../card/card";
 @import "../chart-legend/chart-legend";
 @import "../checkbox/checkbox";
 @import "../chip/chip";
@@ -24,7 +23,9 @@
 @import "../education-notice/education-notice";
 @import "../eek/eek";
 @import "../field/field";
-@import "../file-upload/file-upload";
+@import "../file-preview-card/file-preview-card";
+@import "../file-preview-card-group/file-preview-card-group";
+@import "../file-upload-input/file-upload-input";
 @import "../filter-button/filter-button";
 @import "../filter-menu/filter-menu";
 @import "../filter-menu-button/filter-menu-button";
