+++ src/sass/file-preview-card-group/file-preview-card-group.scss
@@ -1,0 +1,21 @@
+@import "../variables/variables";
+
+div.file-preview-card-group {
+    display: grid;
+    grid-template-columns: repeat(
+        auto-fill,
+        minmax($card-size-default, max-content)
+    );
+    grid-column-gap: var(--spacing-400, 32px);
+    grid-row-gap: var(--spacing-100, 8px);
+}
+
+@media screen and (max-width: $_screen-size-MD) {
+    div.file-upload__assets {
+        grid-template-columns: repeat(
+            auto-fill,
+            minmax($card-size-large, max-content)
+        );
+        grid-column-gap: var(--spacing-100, 8px);
+    }
+}
