+++ src/sass/file-preview-card/file-preview-card.scss
@@ -2,28 +2,28 @@
 @import "../mixins/private/token-mixins";
 @import "../variables/variables";
 
-.card {
+.file-preview-card {
     display: flex;
     flex-direction: column;
     width: $card-size-default;
 }
 
-.card__container {
+.file-preview-card__container {
     position: relative;
     height: $card-size-default;
 }
 
 @media screen and (max-width: $_screen-size-MD) {
-    .card {
+    .file-preview-card {
         width: $card-size-large;
 
-        .card__container {
+        .file-preview-card__container {
             height: $card-size-large;
         }
     }
 }
 
-.card__content {
+.file-preview-card__content {
     display: flex;
     height: inherit;
     width: inherit;
@@ -35,21 +35,21 @@
     padding: var(--spacing-150, 12px);
 }
 
-.card__content .progress-bar {
+.file-preview-card__content .progress-bar {
     width: 100%;
 }
 
-.card__content--image {
+.file-preview-card__content--image {
     background-size: contain;
     background-repeat: no-repeat;
     background-position: center;
 }
 
-.card__content--overflow {
+.file-preview-card__content--overflow {
     opacity: 0.5;
 }
 
-.card .card__action {
+.file-preview-card .file-preview-card__action {
     @include bubble-close();
 
     position: absolute;
@@ -59,7 +59,7 @@
     background: var(--color-background-on-secondary, #fff);
 }
 
-.card__label.icon-btn {
+.file-preview-card__label.icon-btn {
     @include bubble-close();
 
     position: absolute;
@@ -75,13 +75,13 @@
     gap: var(--spacing-50, 4px);
 }
 
-.card__label-text {
+.file-preview-card__label-text {
     color: var(--color-foreground-primary, #191919);
     font-size: var(--font-size-12);
     font-weight: 400;
 }
 
-.card .card__overflow-action {
+.file-preview-card .file-preview-card__overflow-action {
     position: absolute;
     top: calc(50% - 20px);
     left: calc(50% - 20px);
@@ -110,12 +110,12 @@
     }
 }
 
-.card__details {
+.file-preview-card__details {
     display: flex;
     flex-direction: column;
 }
 
-.card__details-title {
+.file-preview-card__details-title {
     color: var(--color-foreground-primary, #191919);
     font-size: var(--font-size-14);
     font-weight: 400;
@@ -123,7 +123,7 @@
     margin-top: var(--spacing-100, 8px);
 }
 
-.card__details-subtitle {
+.file-preview-card__details-subtitle {
     color: var(--color-foreground-secondary, #707070);
     font-size: var(--font-size-12);
     font-weight: 400;
