+++ src/sass/file-upload-input/file-upload-input.scss
@@ -1,32 +1,12 @@
-@import "../variables/variables";
-
-div.file-upload {
-    border-radius: var(--radius-large, 16px);
-}
-
-div.file-upload__assets {
-    display: grid;
-    grid-template-columns: repeat(
-        auto-fill,
-        minmax($card-size-default, max-content)
-    );
-    grid-column-gap: var(--spacing-400, 32px);
-    grid-row-gap: var(--spacing-100, 8px);
-}
-
-div.file-upload__assets--list {
-    grid-template-columns: 1fr;
-}
-
-div.file-uploader {
+div.file-upload-input {
     border-radius: var(--border-radius-100, 16px);
 }
 
-div.file-uploader input.file-uploader-input {
+div.file-upload-input input {
     display: none;
 }
 
-div.file-uploader__container {
+div.file-upload-input__container {
     display: flex;
     flex-direction: row;
     gap: var(--spacing-300, 24px);
@@ -38,7 +18,7 @@ div.file-uploader__container {
     min-width: fit-content;
 }
 
-div.file-uploader__upload-icon-button {
+div.file-upload-input__upload-icon-button {
     background: var(--color-background-on-secondary, #fff);
     border-radius: 99px;
     padding: var(--spacing-300, 24px);
@@ -46,20 +26,20 @@ div.file-uploader__upload-icon-button {
     align-self: center;
 }
 
-div.file-uploader__upload-icon-button svg {
+div.file-upload-input__upload-icon-button svg {
     fill: var(--color-foreground-primary, #191919);
     width: 24px;
     height: 24px;
 }
 
-div.file-uploader__content {
+div.file-upload-input__content {
     display: flex;
     flex-direction: column;
     align-items: start;
     flex-grow: 1;
 }
 
-div.file-uploader__content-header {
+div.file-upload-input__content-header {
     color: var(--color-foreground-primary, #191919);
     font-size: var(--font-size-16);
     font-weight: bold;
@@ -67,7 +47,7 @@ div.file-uploader__content-header {
     margin-bottom: 4px;
 }
 
-div.file-uploader__content-subheader {
+div.file-upload-input__content-subheader {
     color: var(--color-foreground-secondary, #707070);
     font-size: var(--font-size-14);
     font-weight: 400;
@@ -76,35 +56,27 @@ div.file-uploader__content-subheader {
 }
 
 @media screen and (max-width: $_screen-size-MD) {
-    div.file-upload__assets {
-        grid-template-columns: repeat(
-            auto-fill,
-            minmax($card-size-large, max-content)
-        );
-        grid-column-gap: var(--spacing-100, 8px);
-    }
-
-    div.file-uploader__container {
+    div.file-upload-input__container {
         padding: var(--spacing-200, 16px);
         gap: var(--spacing-100, 8px);
     }
 
-    div.file-uploader__upload-icon-button {
+    div.file-upload-input__upload-icon-button {
         padding: var(--spacing-200, 12px);
         max-height: 16px;
     }
 
-    div.file-uploader__upload-icon-button svg {
+    div.file-upload-input__upload-icon-button svg {
         width: 16px;
         height: 16px;
     }
 
-    div.file-uploader__content-header {
+    div.file-upload-input__content-header {
         font-size: var(--font-size-14);
         line-height: 20px;
     }
 
-    div.file-uploader__content-subheader {
+    div.file-upload-input__content-subheader {
         font-size: var(--font-size-12);
         line-height: 16px;
     }
